Fix issues with Lanes and BlockDFromD on SVE
